third_party/blink/web_tests/external/wpt/webmessaging/broadcastchannel/cross-partition.https.tentative-expected.txt

This is a testharness.js-based test.
FAIL BroadcastChannel messages aren't received from a cross-partition iframe assert_equals: expected "msg from iframe2" but got "msg from iframe1"
Harness: the test ran to completion.
